backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
PyMuPDF
requests
python-dotenv
faiss-cpu